+++ manage.py
@@ -1,4 +1,7 @@
+"""Command line tools for Flask server app."""
+
 import unittest
+import coverage
 
 from flask_script import Manager
 
@@ -6,8 +9,44 @@ from app import create_app, db
 from app.api.models import User
 
 
+COV = coverage.coverage(
+    branch=True,
+    include='app/*',
+    omit=[
+        'app/tests/*'
+    ]
+)
+COV.start()
+
+
 app = create_app()
-manager = Manager(app)
+manager = Manager(app)  # pylint: disable=invalid-name
+
+
+@manager.command
+def test():
+    """Run the tests without code coverage."""
+    tests = unittest.TestLoader().discover('app/tests', pattern='test*.py')
+    result = unittest.TextTestRunner(verbosity=2).run(tests)
+    if result.wasSuccessful():
+        return 0
+    return 1
+
+
+@manager.command
+def cov():
+    """Run the unit tests with coverage."""
+    tests = unittest.TestLoader().discover('app/tests')
+    result = unittest.TextTestRunner(verbosity=2).run(tests)
+    if result.wasSuccessful():
+        COV.stop()
+        COV.save()
+        print('Coverage Summary:')
+        COV.report()
+        COV.html_report()
+        COV.erase()
+        return 0
+    return 1
 
 
 @manager.command
@@ -19,13 +58,11 @@ def recreate_db():
 
 
 @manager.command
-def test():
-    """Runs the tests without code coverage."""
-    tests = unittest.TestLoader().discover('app/tests', pattern='test*.py')
-    result = unittest.TextTestRunner(verbosity=2).run(tests)
-    if result.wasSuccessful():
-        return 0
-    return 1
+def seed_db():
+    """Seed the database."""
+    db.session.add(User(username='bchrobot', email="benjamin.blair.chrobot@gmail.com"))
+    db.session.add(User(username='benjaminchrobot', email="benjamin.chrobot@alum.mit.edu"))
+    db.session.commit()
 
 
 if __name__ == '__main__':