+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Basic test suite and commands.
 - 'clean' task to Makefile.
 - Production release process and instructions.
+- Code coverage for testing.
 
 ### Changed
 - Moved API to blueprint.